+++ server/api/transactions/[operation].post.ts
@@ -16,7 +16,7 @@ export default defineEventHandler(async (event) => {
     const user = ensureAuth(event)
 
     // No need to do rest of the logic
-    if(operation === 'delete') {
+    if(operation === 'delete' && id) {
         // Remove the transaction in the database
         await db.deleteFrom('transaction')
             .where('id' , '=', id)
